# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PeerPool decentralized multi-party escrow and dispute-resolution protocol API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type EscrowState = (typeof EscrowState)[keyof typeof EscrowState];

export const EscrowState = {
  pending: "pending",
  funded: "funded",
  active: "active",
  disputed: "disputed",
  settled: "settled",
  closed: "closed",
} as const;

export type DisputeState = (typeof DisputeState)[keyof typeof DisputeState];

export const DisputeState = {
  open: "open",
  escalated: "escalated",
  resolved: "resolved",
  closed: "closed",
} as const;

export type ClaimState = (typeof ClaimState)[keyof typeof ClaimState];

export const ClaimState = {
  pending: "pending",
  submitted: "submitted",
  executed: "executed",
  rejected: "rejected",
} as const;

export type ParticipantRole =
  (typeof ParticipantRole)[keyof typeof ParticipantRole];

export const ParticipantRole = {
  depositor: "depositor",
  beneficiary: "beneficiary",
  arbitrator: "arbitrator",
  observer: "observer",
} as const;

export interface Escrow {
  id: string;
  title: string;
  description?: string;
  state: EscrowState;
  chain: string;
  contractAddress?: string;
  token: string;
  totalAmount: string;
  fundedAmount?: string;
  creatorAddress: string;
  manifestId: string;
  deadline?: string;
  createdAt: string;
  fundedAt?: string;
  participantCount?: number;
  voteCount?: number;
}

export interface EscrowList {
  items: Escrow[];
  total: number;
}

export type EscrowSummaryByState = { [key: string]: number };

export interface EscrowSummary {
  total: number;
  byState: EscrowSummaryByState;
  totalValueLocked: string;
  recentCount: number;
  avgParticipants?: number;
}

export interface CreateEscrowBody {
  title: string;
  description?: string;
  chain: string;
  token: string;
  totalAmount: string;
  creatorAddress: string;
  manifestId: string;
  deadline?: string;
}

export interface UpdateEscrowBody {
  state?: EscrowState;
  contractAddress?: string;
  fundedAmount?: string;
  fundedAt?: string;
}

export interface Participant {
  id: string;
  escrowId: string;
  address: string;
  role: ParticipantRole;
  fundedAmount?: string;
  joinedAt: string;
}

export interface ParticipantList {
  items: Participant[];
}

export interface AddParticipantBody {
  address: string;
  role: ParticipantRole;
  fundedAmount?: string;
}

export interface Vote {
  id: string;
  escrowId: string;
  voterAddress: string;
  outcomeIndex: number;
  outcomeLabel?: string;
  weight?: string;
  signature?: string;
  votedAt: string;
}

export interface OutcomeTally {
  outcomeIndex: number;
  outcomeLabel: string;
  voteCount: number;
  totalWeight: string;
}

export interface VoteList {
  items: Vote[];
  tally: OutcomeTally[];
}

export interface SubmitVoteBody {
  voterAddress: string;
  outcomeIndex: number;
  weight?: string;
  signature?: string;
}

export interface ManifestOutcome {
  index: number;
  label: string;
  description?: string;
  distributionBps?: number;
}

export interface Manifest {
  id: string;
  title: string;
  description?: string;
  conditions: string[];
  outcomes: ManifestOutcome[];
  createdBy: string;
  ipfsHash?: string;
  createdAt: string;
  escrowCount?: number;
}

export interface ManifestList {
  items: Manifest[];
  total: number;
}

export interface CreateManifestBody {
  title: string;
  description?: string;
  conditions: string[];
  outcomes: ManifestOutcome[];
  createdBy: string;
  ipfsHash?: string;
}

export interface Dispute {
  id: string;
  escrowId: string;
  disputerAddress: string;
  reason: string;
  state: DisputeState;
  bondAmount: string;
  klerosDisputeId?: string;
  aiVerdictSummary?: string;
  resolvedOutcomeIndex?: number;
  resolvedBy?: string;
  createdAt: string;
  resolvedAt?: string;
}

export interface DisputeList {
  items: Dispute[];
  total: number;
}

export type DisputeSummaryByState = { [key: string]: number };

export interface DisputeSummary {
  total: number;
  byState: DisputeSummaryByState;
  avgResolutionDays: number;
  klerosEscalationRate: number;
  totalBondsLocked?: string;
}

export interface CreateDisputeBody {
  escrowId: string;
  disputerAddress: string;
  reason: string;
  bondAmount: string;
}

export interface ResolveDisputeBody {
  resolvedOutcomeIndex: number;
  resolvedBy: string;
  aiVerdictSummary?: string;
  klerosDisputeId?: string;
}

export interface Claim {
  id: string;
  escrowId: string;
  claimantAddress: string;
  amount: string;
  merkleRoot?: string;
  merkleProof?: string[];
  leafHash?: string;
  state: ClaimState;
  createdAt: string;
  submittedAt?: string;
  executedAt?: string;
}

export interface ClaimList {
  items: Claim[];
}

export interface CreateClaimBody {
  claimantAddress: string;
  amount: string;
  merkleRoot?: string;
  leafHash?: string;
}

export interface SubmitClaimBody {
  merkleProof: string[];
}

export type ActivityEventType =
  (typeof ActivityEventType)[keyof typeof ActivityEventType];

export const ActivityEventType = {
  escrow_created: "escrow_created",
  escrow_funded: "escrow_funded",
  participant_added: "participant_added",
  vote_submitted: "vote_submitted",
  dispute_opened: "dispute_opened",
  dispute_resolved: "dispute_resolved",
  claim_created: "claim_created",
  claim_executed: "claim_executed",
  manifest_registered: "manifest_registered",
} as const;

export type ActivityEventData = { [key: string]: unknown };

export interface ActivityEvent {
  id: string;
  type: ActivityEventType;
  escrowId: string;
  actorAddress: string;
  data?: ActivityEventData;
  timestamp: string;
}

export interface ActivityList {
  items: ActivityEvent[];
}

export interface ProtocolStats {
  totalEscrows: number;
  totalValueLocked: string;
  totalDisputes: number;
  totalClaims: number;
  totalManifests: number;
  activeChains: string[];
  disputeResolutionRate?: number;
  avgEscrowDurationDays?: number;
}

export type ListEscrowsParams = {
  state?: ListEscrowsState;
  chain?: string;
  limit?: number;
  offset?: number;
};

export type ListEscrowsState =
  (typeof ListEscrowsState)[keyof typeof ListEscrowsState];

export const ListEscrowsState = {
  pending: "pending",
  funded: "funded",
  active: "active",
  disputed: "disputed",
  settled: "settled",
  closed: "closed",
} as const;

export type ListManifestsParams = {
  limit?: number;
  offset?: number;
};

export type ListDisputesParams = {
  state?: ListDisputesState;
  limit?: number;
  offset?: number;
};

export type ListDisputesState =
  (typeof ListDisputesState)[keyof typeof ListDisputesState];

export const ListDisputesState = {
  open: "open",
  escalated: "escalated",
  resolved: "resolved",
  closed: "closed",
} as const;

export type ListActivityParams = {
  escrow_id?: string;
  limit?: number;
};