# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PeerPool decentralized multi-party escrow and dispute-resolution protocol API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /escrows:
    get:
      operationId: listEscrows
      tags: [escrows]
      summary: List all escrows
      parameters:
        - name: state
          in: query
          schema:
            type: string
            enum: [pending, funded, active, disputed, settled, closed]
        - name: chain
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of escrows
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EscrowList"
    post:
      operationId: createEscrow
      tags: [escrows]
      summary: Create a new escrow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEscrowBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Escrow"
...
  /escrows/{id}:
    get:
      operationId: getEscrow
      tags: [escrows]
      summary: Get escrow details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Escrow detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Escrow"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateEscrow
      tags: [escrows]
      summary: Update escrow state
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEscrowBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Escrow"
...
  /escrows/{id}/claims:
    get:
      operationId: listClaims
      tags: [claims]
      summary: List claims for an escrow
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Claims
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClaimList"
    post:
      operationId: createClaim
      tags: [claims]
      summary: Create a settlement claim
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClaimBody"
      responses:
        "201":
          description: Claim created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Claim"
  /escrows/{id}/claims/{claimId}/submit:
    post:
      operationId: submitClaim
      tags: [claims]
      summary: Submit a claim with merkle proof
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: claimId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitClaimBody"
      responses:
        "200":
          description: Claim submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Claim"
...
  /disputes:
    get:
      operationId: listDisputes
      tags: [disputes]
      summary: List all disputes
      parameters:
        - name: state
          in: query
          schema:
            type: string
            enum: [open, escalated, resolved, closed]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Disputes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DisputeList"
    post:
      operationId: createDispute
      tags: [disputes]
      summary: Open a dispute for an escrow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDisputeBody"
      responses:
        "201":
          description: Dispute opened
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dispute"
...
  /disputes/{id}:
    get:
      operationId: getDispute
      tags: [disputes]
      summary: Get dispute details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Dispute
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dispute"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /disputes/{id}/resolve:
    post:
      operationId: resolveDispute
      tags: [disputes]
      summary: Resolve a dispute with a verdict
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResolveDisputeBody"
      responses:
        "200":
          description: Resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dispute"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    EscrowState:
      type: string
      enum: [pending, funded, active, disputed, settled, closed]
    DisputeState:
      type: string
      enum: [open, escalated, resolved, closed]
    ClaimState:
      type: string
      enum: [pending, submitted, executed, rejected]
...
    Escrow:
      type: object
      required: [id, title, state, chain, token, totalAmount, creatorAddress, manifestId, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        state:
          $ref: "#/components/schemas/EscrowState"
        chain:
          type: string
        contractAddress:
          type: string
        token:
          type: string
        totalAmount:
          type: string
        fundedAmount:
          type: string
        creatorAddress:
          type: string
        manifestId:
          type: string
        deadline:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        fundedAt:
          type: string
          format: date-time
        participantCount:
          type: integer
        voteCount:
          type: integer
    EscrowList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Escrow"
        total:
          type: integer
...
    CreateEscrowBody:
      type: object
      required: [title, chain, token, totalAmount, creatorAddress, manifestId]
      properties:
        title:
          type: string
        description:
          type: string
        chain:
          type: string
        token:
          type: string
        totalAmount:
          type: string
        creatorAddress:
          type: string
        manifestId:
          type: string
        deadline:
          type: string
          format: date-time
    UpdateEscrowBody:
      type: object
      properties:
        state:
          $ref: "#/components/schemas/EscrowState"
        contractAddress:
          type: string
        fundedAmount:
          type: string
        fundedAt:
          type: string
          format: date-time
...
    Dispute:
      type: object
      required: [id, escrowId, disputerAddress, reason, state, bondAmount, createdAt]
      properties:
        id:
          type: string
        escrowId:
          type: string
        disputerAddress:
          type: string
        reason:
          type: string
        state:
          $ref: "#/components/schemas/DisputeState"
        bondAmount:
          type: string
        klerosDisputeId:
          type: string
        aiVerdictSummary:
          type: string
        resolvedOutcomeIndex:
          type: integer
        resolvedBy:
          type: string
        createdAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
    DisputeList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Dispute"
        total:
          type: integer
...
    CreateDisputeBody:
      type: object
      required: [escrowId, disputerAddress, reason, bondAmount]
      properties:
        escrowId:
          type: string
        disputerAddress:
          type: string
        reason:
          type: string
        bondAmount:
          type: string
    ResolveDisputeBody:
      type: object
      required: [resolvedOutcomeIndex, resolvedBy]
      properties:
        resolvedOutcomeIndex:
          type: integer
        resolvedBy:
          type: string
        aiVerdictSummary:
          type: string
        klerosDisputeId:
          type: string
    Claim:
      type: object
      required: [id, escrowId, claimantAddress, amount, state, createdAt]
      properties:
        id:
          type: string
        escrowId:
          type: string
        claimantAddress:
          type: string
        amount:
          type: string
        merkleRoot:
          type: string
        merkleProof:
          type: array
          items:
            type: string
        leafHash:
          type: string
        state:
          $ref: "#/components/schemas/ClaimState"
        createdAt:
          type: string
          format: date-time
        submittedAt:
          type: string
          format: date-time
        executedAt:
          type: string
          format: date-time
    ClaimList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Claim"
    CreateClaimBody:
      type: object
      required: [claimantAddress, amount]
      properties:
        claimantAddress:
          type: string
        amount:
          type: string
        merkleRoot:
          type: string
        leafHash:
          type: string
    SubmitClaimBody:
      type: object
      required: [merkleProof]
      properties:
        merkleProof:
          type: array
          items:
            type: string
...