# lib/api-zod/src/generated/types/createEscrowBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PeerPool decentralized multi-party escrow and dispute-resolution protocol API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateEscrowBody {
  title: string;
  description?: string;
  chain: string;
  token: string;
  totalAmount: string;
  creatorAddress: string;
  manifestId: string;
  deadline?: Date;
}