# lib/api-zod/src/generated/types/dispute.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PeerPool decentralized multi-party escrow and dispute-resolution protocol API
 * OpenAPI spec version: 0.1.0
 */
import type { DisputeState } from "./disputeState";

export interface Dispute {
  id: string;
  escrowId: string;
  disputerAddress: string;
  reason: string;
  state: DisputeState;
  bondAmount: string;
  klerosDisputeId?: string;
  aiVerdictSummary?: string;
  resolvedOutcomeIndex?: number;
  resolvedBy?: string;
  createdAt: Date;
  resolvedAt?: Date;
}