# lib/api-zod/src/generated/types/disputeList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PeerPool decentralized multi-party escrow and dispute-resolution protocol API
 * OpenAPI spec version: 0.1.0
 */
import type { Dispute } from "./dispute";

export interface DisputeList {
  items: Dispute[];
  total: number;
}