# lib/api-zod/src/generated/types/escrow.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PeerPool decentralized multi-party escrow and dispute-resolution protocol API
 * OpenAPI spec version: 0.1.0
 */
import type { EscrowState } from "./escrowState";

export interface Escrow {
  id: string;
  title: string;
  description?: string;
  state: EscrowState;
  chain: string;
  contractAddress?: string;
  token: string;
  totalAmount: string;
  fundedAmount?: string;
  creatorAddress: string;
  manifestId: string;
  deadline?: Date;
  createdAt: Date;
  fundedAt?: Date;
  participantCount?: number;
  voteCount?: number;
}